==> 7_similar_dog_demo.ipynb <==
# Data Manipulation
import numpy as np
import pandas as pd

# Visualization
import matplotlib.pyplot as plt
from matplotlib.pyplot import figure
import seaborn as sns
import matplotlib.dates as mdates
from sklearn.preprocessing import StandardScaler

# Machine Learning / Time Series
from scipy import stats
import statsmodels.api as sm
from factor_analyzer import FactorAnalyzer
from IPython.display import Image, display
import dog_func

# Snowflake & ETC
import datetime
import copy

df = pd.read_csv("dog_data_clean.csv", index_col= [0])

personality = ['Affectionate With Family',
       'Good With Young Children', 'Good With Other Dogs',
       'Openness To Strangers', 'Playfulness Level',
       'Watchdog/Protective Nature', 'Adaptability Level',
       'Trainability Level', 'Energy Level', 'Barking Level',
       'Mental Stimulation Needs']

apperance = ['height', 'weight', 'bmi', 'life', 'Shedding Level',
       'Coat Grooming Frequency',  'Coat Length', 'marking']

df_custom = df[apperance]

dog_func.euclidean_by_name(dog_func.add_dist_col(df_custom,'australian-cattle-dog'))

for j,i in enumerate(df.index):
    print("--------------------------")
    print(i)
    df_similiar_dogs= add_dist_col(df_custom,i)
    euclidean_by_name(df_similiar_dogs)
    if i == 50:
        break

# custom
